cmdline.py is for making command-line Python apps.
hello.py is a simple demo.

This is the first commit, mostly written in a late-night hacking spree.

This is also the culmination of some thoughts and work along these lines
over the course of a few years. That history is scattered throughout
various past projects.

So, do not consider this the "beginning". For that, you must seek out
code from a very failed startup I worked on (and ignore some tricky
epistemological questions).