Add basic support for option/flag short names.
Also introduces some complexity to the App.command() and App.main()
decorators, to avoid adding duplicate decorators with different names
that handle the rather different semantics Python uses for decorators
that are passed args.

As a result, these decorators may only be passed keyword args.

That's arguably better for readability anyway, but it's still a wart. It
is not Pythonic to *force* you to write good code.